requirements.txt
jax
pgx
mctx
optax
flax
hydra-core
chex
chess
numpy
omegaconf
etils
tqdm
grain
tensorflow-datasets